More changes from xpilot-ng, including x11 client optionally painting direction pointer.

+++ src/server/collision.cpp
@@ -606,7 +606,7 @@ static void PlayerObjectCollision(player_t *pl)
 {
     world_t *world = &World;
     int j, obj_count;
-    int range, radius;
+    double range, radius;
     object_t *obj, **obj_list;
 
     /*
@@ -1043,6 +1043,7 @@ static void Player_collides_with_mine(player_t *pl, mineobject_t *mine)
     else if (mine->mine_owner == NO_ID)
     {
         const char *reprogrammer_name = "some jerk";
+
         if (mine->id != NO_ID)
         {
             kp = Player_by_id(mine->id);
@@ -1056,6 +1057,7 @@ static void Player_collides_with_mine(player_t *pl, mineobject_t *mine)
     else
     {
         const char *reprogrammer_name = "some jerk";
+
         if (mine->id != NO_ID)
         {
             kp = Player_by_id(mine->id);
@@ -1145,7 +1147,7 @@ static void Player_collides_with_asteroid(player_t *pl, wireobject_t *ast)
     if (!Player_uses_emergency_shield(pl))
         Player_add_fuel(pl, -cost);
 
-    if (options.asteroidCollisionMayKill && (pl->fuel.sum == 0 || (!BIT(pl->used, HAS_SHIELD) && !Player_has_armor(pl))))
+    if (options.asteroidCollisionMayKill && (pl->fuel.sum == 0.0 || (!BIT(pl->used, HAS_SHIELD) && !Player_has_armor(pl))))
     {
         int sc;
         Player_set_state(pl, PL_STATE_KILLED);
